Issue 65: Android: use scalable pixels for fonts instead of "real" pixels
git-svn-id: http://microemu.googlecode.com/svn/trunk@2419 272dc960-a678-11de-8dbb-b5444cbab91d

+++ microemulator/microemu-android/src/org/microemu/android/device/AndroidFont.java
@@ -38,7 +38,7 @@ public class AndroidFont implements Font {
 	
 	public FontMetricsInt metrics;
 	
-	public AndroidFont(Typeface typeface, int size, boolean underlined) {
+	public AndroidFont(Typeface typeface, float size, boolean underlined) {
 		paint.setTypeface(typeface);
 		paint.setTextSize(size);
 		paint.setUnderlineText(underlined);
